direct access  onGetMediaActionControl
