+++ colan_services/lib/colan_services.dart
@@ -9,9 +9,12 @@ export 'services/incoming_media_service/incoming_media_monitor.dart'
     show IncomingMediaMonitor;
 export 'services/incoming_media_service/incoming_media_service.dart';
 export 'services/incoming_media_service/models/cl_shared_media.dart';
+export 'services/media_edit_service/media_edit_service.dart'
+    show MediaEditService;
 export 'services/media_view_service/collection_view.dart' show CollectionView;
 export 'services/media_view_service/media_view_service.dart'
     show MediaViewService;
+export 'services/media_view_service/models/action_control.dart';
 export 'services/media_view_service/widgets/media_gallery_view.dart';
 export 'services/media_view_service/widgets/stale_media_indicator.dart';
 export 'services/media_wizard_service/media_wizard_service.dart';
@@ -20,6 +23,5 @@ export 'services/notification_services/notification_service.dart'
     show NotificationService;
 export 'services/notification_services/provider/notify.dart'
     show notificationMessageProvider;
-
 export 'services/quick_menu_service/collection_menu.dart' show CollectionMenu;
 export 'services/quick_menu_service/media_menu.dart' show MediaMenu;
