+++ colan_services/lib/services/media_view_service/models/action_control.dart
@@ -1,0 +1,37 @@
+import 'package:colan_widgets/colan_widgets.dart';
+import 'package:media_editors/media_editors.dart';
+import 'package:store/store.dart';
+
+extension AccessControlExt on ActionControl {
+  static ActionControl actionControlNone() {
+    return const ActionControl(
+      allowEdit: false,
+      allowDelete: false,
+      allowMove: false,
+      allowShare: false,
+      allowPin: false,
+      canDuplicateMedia: false,
+    );
+  }
+
+  static ActionControl onGetMediaActionControl(CLMedia media) {
+    final editSupported = switch (media.type) {
+      CLMediaType.text => false,
+      CLMediaType.image => true,
+      CLMediaType.video => VideoEditor.isSupported,
+      CLMediaType.url => false,
+      CLMediaType.audio => false,
+      CLMediaType.file => false,
+    };
+    final isAFile = media.isMediaCached;
+
+    return ActionControl(
+      allowEdit: editSupported && isAFile,
+      allowDelete: true,
+      allowMove: true,
+      allowShare: isAFile,
+      allowPin: isAFile,
+      canDuplicateMedia: true,
+    );
+  }
+}
